# Performance Improvements and UI Overhaul for Load Testing