--- a/.tool-versions
+++ b/.tool-versions
@@ -1,0 +1,1 @@
+rust stable