--- a/dev/dashboard/Dockerfile
+++ b/dev/dashboard/Dockerfile
@@ -1,7 +1,0 @@
-FROM node:18-alpine
-WORKDIR /app
-COPY package*.json ./
-RUN npm install
-COPY . .
-EXPOSE 5173
-CMD ["npm", "run", "dev", "--", "--host"]