--- a/dev/dashboard/src/main.jsx
+++ b/dev/dashboard/src/main.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import App from './App.jsx';
-
-ReactDOM.createRoot(document.getElementById('root')).render(<App />);