--- a/monitoring/prometheus/prometheus.yml
+++ b/monitoring/prometheus/prometheus.yml
@@ -13,6 +13,6 @@ scrape_configs:
 
   - job_name: 'banking-api'
     static_configs:
-      - targets: ['api:8080']
+      - targets: ['banking-api:8080']
     metrics_path: '/prometheus'
     scrape_interval: 5s