--- a/perf-test/go.mod
+++ b/perf-test/go.mod
@@ -1,0 +1,10 @@
+module github.com/core-banking/perf-test
+
+go 1.21
+
+require (
+	github.com/gorilla/mux v1.8.1
+	github.com/gorilla/websocket v1.5.1
+)
+
+require golang.org/x/net v0.17.0 // indirect