… CLAUDE.md
… PERFORMANCE_OPTIMIZATION.md
… dev/dashboard/index.html
… dev/dashboard/package.json
… dev/dashboard/src/App.jsx
… dev/dashboard/src/simulator.js
… dev/dashboard/src/styles.css
… dev/dashboard/vite.config.js
… docker-compose.yml
… k8s/01-banking-api-optimized.yaml
… k8s/02-dashboard-deployment.yaml
… k8s/02-perf-test-deployment.yaml
… monitoring/grafana/dashboards/banking-overview.json
… monitoring/grafana/dashboards/kubernetes-monitoring.json
… perf-test/README.md
… perf-test/docker-compose.yml
… perf-test/go.sum
… perf-test/internal/metrics/collector.go
… perf-test/internal/metrics/prometheus.go
… perf-test/internal/monitor/system.go
… perf-test/internal/reporter/reporter.go
… perf-test/internal/server/server.go
… perf-test/reports/report_1756057059.json
… perf-test/reports/report_1756057094.json
… perf-test/reports/report_1756057308.json
… perf-test/reports/report_1756057893.json
… perf-test/reports/report_1756058230.json
… perf-test/reports/report_1756058483.json
… perf-test/reports/report_1756058747.json
… perf-test/reports/report_test-1756057363.json