## Summary

This PR introduces comprehensive performance optimizations and a complete UI overhaul for the Core Banking load testing system. The changes address performance bottlenecks and provide a modern, feature-rich testing interface.

### 🚀 Performance Improvements (6.2K → 50K+ RPS target)
- **Banking API Optimizations**: GOMAXPROCS=8, GOGC=400, 2GB memory limits
- **Runtime Tuning**: Memory ballast, reduced GC frequency, multi-core utilization  
- **Container Optimization**: Multi-stage Docker builds, Alpine base images
- **Resource Allocation**: 8 CPU cores, proper Kubernetes resource management

### 🎨 UI/UX Overhaul
- **New Performance Testing UI**: Modern Bootstrap-based interface
- **Brazilian Localization**: R$ 1.000,00 currency formatting, Portuguese labels
- **Real-time Monitoring**: Live charts, WebSocket updates, performance metrics
- **Enhanced Usability**: Tooltips, input validation, responsive design

### 🔧 Technical Improvements  
- **Test Lifecycle Management**: Fixed start/stop race conditions, auto-cleanup
- **Monitoring Integration**: Prometheus metrics, Grafana dashboard fixes
- **Code Organization**: Modular Go backend, clean separation of concerns
- **Documentation**: Comprehensive performance optimization guide

### 🧹 Cleanup
- **Removed Deprecated Components**: React dashboard, unused Docker services
- **Updated Configuration**: Grafana dashboards, Prometheus scraping
- **Improved Developer Experience**: Updated CLAUDE.md, version management

## Test Plan

- [x] Banking API serves requests with optimized performance
- [x] Performance test UI loads and displays metrics correctly
- [x] Start/stop test cycle works reliably with different parameters
- [x] Grafana dashboards show accurate CPU/memory metrics (~6% not 100%)
- [x] Brazilian currency formatting displays correctly
- [x] Test completion triggers automatic cleanup
- [x] Kubernetes deployments succeed with proper resource allocation

## Expected Performance Impact

| Metric | Before | After (Target) |
|--------|--------|----------------|
| **Throughput** | 6.2K RPS | 50K+ RPS |
| **CPU Usage** | 95% (1 core) | 40-60% (8 cores) |
| **GC Pause** | 950ms | <50ms |
| **P99 Latency** | 100ms | <10ms |

🤖 Generated with [Claude Code](https://claude.ai/code)